+++ libraries/transformer/src/main/java/androidx/media3/transformer/Muxer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package androidx.media3.muxer;
+package androidx.media3.transformer;
 
 import android.media.MediaCodec.BufferInfo;
 import androidx.media3.common.C;
@@ -21,6 +21,7 @@ import androidx.media3.common.Format;
 import androidx.media3.common.Metadata;
 import androidx.media3.common.MimeTypes;
 import androidx.media3.common.util.UnstableApi;
+import androidx.media3.muxer.MuxerException;
 import com.google.common.collect.ImmutableList;
 import java.nio.ByteBuffer;