+++ RELEASENOTES.md
@@ -37,6 +37,7 @@
         `InAppMp4Muxer` is to be used for producing a non-fragmented MP4 file,
         while `InAppFragmentedMp4Muxer` is to be used for producing a fragmented
         MP4 file.
+    *   Move `Muxer` interface from `media3-muxer` to `media3-transformer`.
 *   Track Selection:
 *   Extractors:
     *   Fix handling of NAL units with lengths expressed in 1 or 2 bytes (rather